new factory folder and moved service functions
